+++ src/controllers/recepcaoEvento-controller.js
@@ -1,42 +1,11 @@
 "use strict"
 
-const https = require("https")
-
 const schemaRecepcaoEvento = require("../schema/recepcaoEvento")
 const Client = require("../services/client-service")
 const { assinaturaXml, enveloparXml, json2xml, xml2json } = require("../util")
 
-const URL = {
-  1: "https://www.nfe.fazenda.gov.br/NFeRecepcaoEvento4/NFeRecepcaoEvento4.asmx?wsdl",
-  2: "https://hom.nfe.fazenda.gov.br/NFeRecepcaoEvento4/NFeRecepcaoEvento4.asmx?wsdl",
-}
-
 exports.enviar = async (opts) => {
   const { requestOpt, httpsOpt } = opts
-  const baseURL = URL[opts.tpAmb]
-
-  const AgentOptions = Object.assign(
-    {
-      cert: opts.cert,
-      key: opts.key,
-    },
-    { ...httpsOpt }
-  )
-
-  const httpsAgent = new https.Agent(AgentOptions)
-
-  const requestOptions = Object.assign(
-    {
-      baseURL: baseURL,
-      headers: {
-        "User-Agent": `node-mde`,
-        "Content-Type": "text/xml;charset=utf-8",
-      },
-      httpsAgent: httpsAgent,
-      timeout: 60000,
-    },
-    { ...requestOpt }
-  )
 
   const schema = schemaRecepcaoEvento.schema(opts)
   const xml = json2xml(schema)
@@ -48,7 +17,12 @@ exports.enviar = async (opts) => {
     data: data,
   }
 
-  const client = new Client({ requestOptions: requestOptions })
+  const client = new Client({
+    service: "recepcao",
+    tpAmb: opts.tpAmb,
+    requestOptions: requestOpt,
+    httpsOptions: httpsOpt,
+  })
 
   const eventoRetorno = await client.request(options)
 
@@ -64,54 +38,55 @@ exports.enviar = async (opts) => {
 }
 
 /**
- * @returns {{retEnvEvento:{idLote: string,tpAmb: string,verAplic: string,cOrgao: string,cStat: string,xMotivo: string}, infEvento:{tpAmb: string,verAplic: string,cOrgao: string,cStat: string,xMotivo: string,chNFe: string,tpEvento: string,xEvento: string,nSeqEvento: string,dhRegEvento: string}} | {error: string}}
+ * @returns {{retEnvEvento:{idLote: string,tpAmb: string,verAplic: string,cOrgao: string,cStat: string,xMotivo: string}, infEvento:{tpAmb: string,verAplic: string,cOrgao: string,cStat: string,xMotivo: string,chNFe: string,tpEvento: string,xEvento: string,nSeqEvento: string,dhRegEvento: string}, error: string}}
  */
 function montarRetorno(data) {
-  try {
-    const json = xml2json(data)
+  const retorno = {
+    retEnvEvento: {},
+    infEvento: {},
+    error: "",
+  }
 
-    if (json.error) {
-      throw new Error(json?.error?.value || "Falha ao montar retorno do SEFAZ.")
-    }
+  const json = xml2json(data)
 
-    const {
-      "soap:Envelope": {
-        "soap:Body": {
-          nfeRecepcaoEventoNFResult: {
-            retEnvEvento: {
-              retEvento: { infEvento = {} } = {},
-              ...retEnvEvento
-            } = {},
+  if (json.error) {
+    retorno["error"] = json?.error?.value || "Falha ao montar retorno do SEFAZ."
+  }
+
+  const {
+    "soap:Envelope": {
+      "soap:Body": {
+        nfeRecepcaoEventoNFResult: {
+          retEnvEvento: {
+            retEvento: { infEvento = {} } = {},
+            ...retEnvEvento
           } = {},
         } = {},
       } = {},
-    } = json
+    } = {},
+  } = json
 
-    const retorno = {
-      retEnvEvento: {
-        idLote: retEnvEvento?.idLote?.value || "",
-        tpAmb: retEnvEvento?.tpAmb?.value || "",
-        verAplic: retEnvEvento?.verAplic?.value || "",
-        cOrgao: retEnvEvento?.cOrgao?.value || "",
-        cStat: retEnvEvento?.cStat?.value || "",
-        xMotivo: retEnvEvento?.xMotivo?.value || "",
-      },
-      infEvento: {
-        tpAmb: infEvento?.tpAmb?.value || "",
-        verAplic: infEvento?.verAplic?.value || "",
-        cOrgao: infEvento?.cOrgao?.value || "",
-        cStat: infEvento?.cStat?.value || "",
-        xMotivo: infEvento?.xMotivo?.value || "",
-        chNFe: infEvento?.chNFe?.value || "",
-        tpEvento: infEvento?.tpEvento?.value || "",
-        xEvento: infEvento?.xEvento?.value || "",
-        nSeqEvento: infEvento?.nSeqEvento?.value || "",
-        dhRegEvento: infEvento?.dhRegEvento?.value || "",
-      },
-    }
-
-    return retorno
-  } catch (e) {
-    return { error: `${e.message || e}` }
+  retorno["retEnvEvento"] = {
+    idLote: retEnvEvento?.idLote?.value || "",
+    tpAmb: retEnvEvento?.tpAmb?.value || "",
+    verAplic: retEnvEvento?.verAplic?.value || "",
+    cOrgao: retEnvEvento?.cOrgao?.value || "",
+    cStat: retEnvEvento?.cStat?.value || "",
+    xMotivo: retEnvEvento?.xMotivo?.value || "",
   }
+
+  retorno["infEvento"] = {
+    tpAmb: infEvento?.tpAmb?.value || "",
+    verAplic: infEvento?.verAplic?.value || "",
+    cOrgao: infEvento?.cOrgao?.value || "",
+    cStat: infEvento?.cStat?.value || "",
+    xMotivo: infEvento?.xMotivo?.value || "",
+    chNFe: infEvento?.chNFe?.value || "",
+    tpEvento: infEvento?.tpEvento?.value || "",
+    xEvento: infEvento?.xEvento?.value || "",
+    nSeqEvento: infEvento?.nSeqEvento?.value || "",
+    dhRegEvento: infEvento?.dhRegEvento?.value || "",
+  }
+
+  return retorno
 }