+++ src/services/client-service.js
@@ -1,10 +1,44 @@
 "use strict"
 
 const axios = require("axios").default
+const https = require("https")
 
+const URL = {
+  distribuicao: {
+    1: "https://www1.nfe.fazenda.gov.br/NFeDistribuicaoDFe/NFeDistribuicaoDFe.asmx?wsdl",
+    2: "https://hom1.nfe.fazenda.gov.br/NFeDistribuicaoDFe/NFeDistribuicaoDFe.asmx?wsdl",
+  },
+  recepcao: {
+    1: "https://www.nfe.fazenda.gov.br/NFeRecepcaoEvento4/NFeRecepcaoEvento4.asmx?wsdl",
+    2: "https://hom.nfe.fazenda.gov.br/NFeRecepcaoEvento4/NFeRecepcaoEvento4.asmx?wsdl",
+  },
+}
 class Instance {
-  constructor(options) {
-    const requestOptions = options.requestOptions
+  constructor(opts) {
+    const baseURL = URL[opts.service][opts.tpAmb]
+
+    const AgentOptions = Object.assign(
+      {
+        cert: opts.cert,
+        key: opts.key,
+      },
+      { ...opts.httpsOptions }
+    )
+
+    const httpsAgent = new https.Agent(AgentOptions)
+
+    const requestOptions = Object.assign(
+      {
+        baseURL: baseURL,
+        headers: {
+          "User-Agent": `node-mde`,
+          "Content-Type": "text/xml;charset=utf-8",
+        },
+        httpsAgent: httpsAgent,
+        timeout: 60000,
+      },
+      { ...opts.requestOptions }
+    )
 
     const instance = axios.create({
       ...requestOptions,
@@ -14,7 +48,7 @@ class Instance {
   }
 
   /**
-   * @returns {{status: number, data: string}}
+   * @returns {Promise<{status: number, data: string}>}
    */
   async request(options) {
     try {